services/graphql-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
strawberry-graphql[fastapi]==0.214.0
prometheus-fastapi-instrumentator==7.0.0
